requirements.txt
discord.py
python-dotenv
googletrans==4.0.0rc1
requests
yt-dlp